requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
diffusers>=0.29.0
accelerate>=0.30.0
safetensors>=0.4.2
numpy
Pillow
matplotlib
tqdm
